Updated handling of RegionOfInterest in post processing techniques
Previously attempting to perform PCA (or other MVA) technique on a ROI would fail. If the ROI had different pixels to those present within the data, this would also fail. Now the correct data is extracted for PCA (or other MVA) and a new (correct) ROI is generated from the combination of the original data ROI and the user selected ROI.